# docs/swagger.yaml
basePath: /
definitions:
...
  admin.createMovieRequest:
    properties:
      ages:
        items:
          type: integer
        type: array
      categories:
        items:
          type: integer
        type: array
      description:
        type: string
      director:
        type: string
      genres:
        items:
          type: integer
        type: array
      keywords:
        items:
          type: string
        type: array
      movieTypeId:
        type: integer
      producer:
        type: string
      releaseYear:
        type: integer
      runtime:
        type: integer
      title:
        type: string
    type: object
...
  models.Ages:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Category:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
  models.Episode:
    properties:
      id:
        type: integer
      number:
        description: Номер серии
        type: integer
      seasonID:
        description: ID сезона, к которому относится эпизод
        type: integer
      videoURL:
        description: Ссылка на видео
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
    type: object
  models.Movie:
    properties:
      ages:
        items:
          $ref: '#/definitions/models.Ages'
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      cover:
        type: string
      description:
        type: string
      director:
        type: string
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      keyWords:
        items:
          type: string
        type: array
      movieType:
        type: string
      movieTypeId:
        type: integer
      producer:
        type: string
      releaseYear:
        type: integer
      runtime:
        type: integer
      screenshots:
        items:
          type: string
        type: array
      seasons:
        items:
          $ref: '#/definitions/models.Season'
        type: array
      title:
        type: string
    type: object
...
  models.Season:
    properties:
      episodes:
        description: Связь один ко многим
        items:
          $ref: '#/definitions/models.Episode'
        type: array
      id:
        type: integer
      movieID:
        description: ID фильма, к которому относится сезон
        type: integer
      number:
        description: Номер сезона
        type: integer
    type: object
...
host: localhost:8081
info:
  contact:
    email: bexultanovmadiyar@gmail.com
    name: Madiyar Bexultanov
    url: https://www.linkedin.com/in/madiyar-bexultanov-b21902258/
  description: Personal online platform providing information about media content
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ozinshe Production
  version: "1.0"
paths:
...
  /admin/movies:
    get:
      consumes:
      - application/json
      description: Get a list of all movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all movies
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Create a new movie with the specified information
      parameters:
      - description: Movie information
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/admin.createMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Movie ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create a new movie
      tags:
      - Movies
  /admin/movies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a movie by its ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Movie deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete movie
      tags:
      - Movies
    get:
      consumes:
      - application/json
      description: Get a movie by its ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get movie by ID
      tags:
      - Movies
    patch:
      consumes:
      - application/json
      description: Update an existing movie by its ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Movie information
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/admin.createMovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Movie updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update movie
      tags:
      - Movies
  /admin/movies/{movieId}/media:
    patch:
      consumes:
      - multipart/form-data
      description: Upload cover and screenshots for a movie
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      - description: Cover image
        in: formData
        name: cover
        required: true
        type: file
      - description: Screenshots
        in: formData
        name: screenshots
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Media added successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Add media (cover, screenshots)
      tags:
      - Movies
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"